Move away from replaceWebpackConfig
Moved away from `replaceWebpackConfig` in favor of a less intrusive
approach using `setWebpackConfig`. This is useful because plugins in
the Gatsby ecosystem (such as `gatsby-mdx`) have access to modify the
webpack config as well, and `setWebpackConfig` composes configuration
well, while `replaceWebpackConfig` causes the user to need to take
account of every plugin that could be adding additional configuration.

Also fixed a path import in an MDX file.